working, todo add example cfg
